Generalize input-blocking mod for taiko
